+++ javascript/modules/client/CHANGELOG.md
@@ -18,6 +18,9 @@ TEMPLATE:
 
 ## [UPCOMING]
 
+## 0.0.16-alpha
+On 2022-06-14 15:03:15
+
 - Gas fee estimation for `deposit` and `increaseAllowance`.
 
 ### Changed